+++ jest.config.js
@@ -1,0 +1,19 @@
+module.exports = {
+  verbose: true,
+  roots: ['<rootDir>/src'],
+  moduleNameMapper: {
+    '\\.(css|less|scss)$': 'identity-obj-proxy',
+    // '^components$': '<rootDir>/components/index.tsx',
+    // '^components(.*)$': '<rootDir>/components/$1',
+  },
+  testRegex: '(/test/.*|\\.(test|spec))\\.(ts|tsx|js)$',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  testPathIgnorePatterns: ['/node_modules/', '/lib/', '/esm/', '/dist/'],
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  transform: {
+    "^.+\\.(js|ts|tsx|jsx)$": "babel-jest",
+    ".+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$": "jest-transform-stub"
+  },
+  setupFilesAfterEnv: ['<rootDir>/scripts/setupTests.js'],
+};