+++ src/button/index.ts
@@ -1,0 +1,9 @@
+/**
+ * @FileName src/button/index.ts
+ * @Description Button interface
+ * @Author qulongjun <contact@nuwa.design>
+ * @Date 2021/11/26 16:16:12
+ */
+export * from './interface';
+
+export { default } from './button';