+++ src/common/interfaces/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @FileName src/common/interfaces/index.ts
+ * @Description 公共接口 入口
+ * @Author qulongjun <contact@nuwa.design>
+ * @Date 2021/11/26 15:59:11
+ */
+
+// 组件通用 Props Interface
+export * from './ICommonProps';
+
+// 全局配置通用 Config Interface
+export * from './IGlobalConfig';